# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: JARVIS Backend API
  version: 1.1.0
  description: |
    v1 API contract aligned with currently implemented Fastify routes.
servers:
  - url: https://api.jarvis.local
    description: Default server
...
paths:
...
  /api/v1/radar/ingest:
    post:
      tags: [Radar]
      summary: Trigger radar ingestion pipeline
      operationId: ingestRadar
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RadarIngestRequest'
      responses:
        '202':
          description: Ingestion accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessRadarIngest'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/radar/items:
    get:
      tags: [Radar]
      summary: List normalized radar items
      operationId: listRadarItems
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [new, scored, archived]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Radar item list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessRadarItems'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/radar/evaluate:
    post:
      tags: [Radar]
      summary: Evaluate radar items and generate recommendation scores
      operationId: evaluateRadar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [item_ids]
              properties:
                item_ids:
                  type: array
                  minItems: 1
                  items:
                    type: string
      responses:
        '202':
          description: Evaluation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessRadarEvaluation'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/radar/recommendations:
    get:
      tags: [Radar]
      summary: Get scored recommendations
      operationId: getRadarRecommendations
      parameters:
        - in: query
          name: decision
          schema:
            type: string
            enum: [adopt, hold, discard]
      responses:
        '200':
          description: Recommendation list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessRadarRecommendations'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/radar/reports/telegram:
    get:
      tags: [Radar]
      summary: List telegram report queue/history
      operationId: listRadarTelegramReports
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [queued, sent, failed]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 30
      responses:
        '200':
          description: Telegram report list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessTelegramReports'
        default:
          $ref: '#/components/responses/ApiError'
    post:
      tags: [Radar]
      summary: Send radar digest to Telegram
      operationId: sendRadarTelegramReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chat_id]
              properties:
                chat_id:
                  type: string
      responses:
        '202':
          description: Report enqueue accepted (meta may include signed approval_action_payloads)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessTelegramReport'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/radar/reports/telegram/{reportId}:
    get:
      tags: [Radar]
      summary: Get single telegram report status
      operationId: getRadarTelegramReport
      parameters:
        - in: path
          name: reportId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Telegram report detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessTelegramReport'
        '404':
          $ref: '#/components/responses/ApiError'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/radar/reports/telegram/{reportId}/retry:
    post:
      tags: [Radar]
      summary: Retry failed telegram report delivery
      operationId: retryRadarTelegramReport
      parameters:
        - in: path
          name: reportId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                max_attempts:
                  type: integer
                  minimum: 1
                  maximum: 20
      responses:
        '202':
          description: Retry queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessTelegramReport'
        '404':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/radar/reports/telegram/events:
    get:
      tags: [Radar]
      summary: Stream telegram report list updates
      operationId: streamRadarTelegramReports
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [queued, sent, failed]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 30
      responses:
        '200':
          description: 'text/event-stream payload (event: stream.open|telegram.reports.updated|stream.close)'
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/radar/reports/telegram/{reportId}/events:
    get:
      tags: [Radar]
      summary: Stream single telegram report updates
      operationId: streamRadarTelegramReport
      parameters:
        - in: path
          name: reportId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'text/event-stream payload (event: stream.open|telegram.report.updated|stream.close)'
          content:
            text/event-stream:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/ApiError'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/upgrades/proposals:
    get:
      tags: [Upgrades]
      summary: List upgrade proposals
      operationId: listUpgradeProposals
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [proposed, approved, planning, running, verifying, deployed, failed, rolled_back, rejected]
      responses:
        '200':
          description: Proposal list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessUpgradeProposals'
        default:
          $ref: '#/components/responses/ApiError'
...
components:
  responses:
    ApiError:
      description: Standard API error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
  schemas:
    ApiMeta:
      type: object
      additionalProperties: true
    ApiErrorEnvelope:
      type: object
      required: [request_id, error]
      properties:
        request_id:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    RadarIngestItem:
      type: object
      required: [title, source_url]
      properties:
        title:
          type: string
        summary:
          type: string
        source_url:
          type: string
          format: uri
        published_at:
          type: string
          format: date-time
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
    RadarIngestRequest:
      type: object
      properties:
        source_name:
          type: string
          default: manual
        items:
          type: array
          items:
            $ref: '#/components/schemas/RadarIngestItem'
    RadarItem:
      type: object
      required: [id, title, summary, sourceUrl, sourceName, confidenceScore, status]
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        sourceUrl:
          type: string
          format: uri
        sourceName:
          type: string
        publishedAt:
          type: string
          format: date-time
          nullable: true
        confidenceScore:
          type: number
          minimum: 0
          maximum: 1
        status:
          type: string
          enum: [new, scored, archived]
    RadarRecommendation:
      type: object
      required: [id, itemId, decision, totalScore, expectedBenefit, migrationCost, riskLevel, evaluatedAt]
      properties:
        id:
          type: string
        itemId:
          type: string
        decision:
          type: string
          enum: [adopt, hold, discard]
        totalScore:
          type: number
        expectedBenefit:
          type: string
        migrationCost:
          type: string
        riskLevel:
          type: string
        evaluatedAt:
          type: string
          format: date-time
    UpgradeProposal:
      type: object
      required: [id, recommendationId, proposalTitle, status, createdAt]
      properties:
        id:
          type: string
        recommendationId:
          type: string
        proposalTitle:
          type: string
        status:
          type: string
          enum: [proposed, approved, planning, running, verifying, deployed, failed, rolled_back, rejected]
        createdAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
          nullable: true
...
    TelegramReport:
      type: object
      required: [id, chatId, status, createdAt]
      properties:
        id:
          type: string
        chatId:
          type: string
        topic:
          type: string
        bodyMarkdown:
          type: string
        status:
          type: string
          enum: [queued, sent, failed]
        attemptCount:
          type: integer
        maxAttempts:
          type: integer
        nextAttemptAt:
          type: string
          format: date-time
          nullable: true
        lastError:
          type: string
          nullable: true
        telegramMessageId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        sentAt:
          type: string
          format: date-time
          nullable: true
    TelegramInlineKeyboardButton:
      type: object
      required: [text, callback_data]
      properties:
        text:
          type: string
        callback_data:
          type: string
          description: Signed approval callback token (`tcb1|<action_code>|<proposal_id>|<expires_at_sec>|<nonce_hex>|<sig>`).
    TelegramReplyMarkup:
      type: object
      required: [inline_keyboard]
      properties:
        inline_keyboard:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/TelegramInlineKeyboardButton'
    TelegramApprovalActionPayload:
      type: object
      required: [proposal_id, reply_markup]
      properties:
        proposal_id:
          type: string
          format: uuid
        reply_markup:
          $ref: '#/components/schemas/TelegramReplyMarkup'
    TelegramReportMeta:
      type: object
      properties:
        approval_action_payloads:
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/TelegramApprovalActionPayload'
        telegram_delivery:
          type: object
          required: [attempted, delivered]
          properties:
            attempted:
              type: boolean
            delivered:
              type: boolean
            message_id:
              type: string
              nullable: true
            reason:
              type: string
      additionalProperties: true
...
    ApiSuccessRadarIngest:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [ingest_job_id, status, accepted_count]
          properties:
            ingest_job_id:
              type: string
            status:
              type: string
              enum: [queued]
            accepted_count:
              type: integer
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessRadarItems:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/RadarItem'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessRadarEvaluation:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [evaluation_job_id, status, recommendation_count]
          properties:
            evaluation_job_id:
              type: string
            status:
              type: string
              enum: [queued]
            recommendation_count:
              type: integer
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessRadarRecommendations:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [recommendations]
          properties:
            recommendations:
              type: array
              items:
                $ref: '#/components/schemas/RadarRecommendation'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessTelegramReport:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          $ref: '#/components/schemas/TelegramReport'
        meta:
          $ref: '#/components/schemas/TelegramReportMeta'
    ApiSuccessTelegramReports:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [reports]
          properties:
            reports:
              type: array
              items:
                $ref: '#/components/schemas/TelegramReport'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessUpgradeProposals:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [proposals]
          properties:
            proposals:
              type: array
              items:
                $ref: '#/components/schemas/UpgradeProposal'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...